BatteryService: Use Settings app to enable battery polling
An item has been added to the Settings app so that user can
enable/disable battery polling if they so desire.

The battery level polling is enabled by an entry in the
settings database, instead of system properties. The Settings
app cannot write system properties due to the property
service refusing to honor the request.

For: AIA-432
Change-Id: Idf44606ba3a566341bf5952b21edce3b4101f457
Signed-off-by: Daniel Leung <daniel.leung@intel.com>