BatteryService: added a Timer task to poll battery level
BatteryService in Android depends on uevent to update battery level.
This requires the embedded controller to generate interrupts or GPE
(General Purpose Event).

For some devices, such as Lenovo ThinkPad X220 Tablet, GPEs are never
generated when battery level changes. Due to this, the battery level
is never updated unless some other events (e.g. plugging/unplugging
AC adapter) are triggered.

For this reason, we need to poll the battery level. This patch sets
up a Timer task to poll the status every minute. The timer itself
is usually suspended when device sleeps, so it will not cause
unnecessary wakeup just to check battery level.

The BatteryService only polls the battery level if the persistent
property persist.svc.battery.poll_ms has non-zero positive value
(which is in ms).

For: AIA-432
Change-Id: I8e6f1549f68ca75e203f0ab66d789c93e67f1770
Signed-off-by: Daniel Leung <daniel.leung@intel.com>