Add power off button to status bar pop-up
Add an option to power off the device in the status bar pop-up
dialog. It is necessary for device which the power button does
not function as Android expected. For example, the hardware
only sends button down event but no button up event, or there
is not long press event emitted what-so-ever. Clicking on this
button asks user for confirmation, and fires the shutdown
intent.

The theme for ShutdownActivity is also changed so that it only
displays the dialog, or else it displays a blank window which
looks inconsistent with the power off behavior. Effectively,
the change in theme makes the window translucent. This should
provide a more pleasing look for other system applications and
services too.

Issue: AXIA-76
Change-Id: Ic973ebf43b79b436a9e872613b8572a7c77ce0c3
Signed-off-by: Daniel Leung <daniel.leung@intel.com>
Signed-off-by: Anton Cherkashyn <antonx.t.cherkashyn@intel.com>

Conflicts:

	packages/SystemUI/AndroidManifest.xml
	packages/SystemUI/res/layout-sw600dp/status_bar_settings_view.xml