Add canDetectOrientation to WM & RotationPolicy
Expose WindowOrientationListener's canDetectOrientation method up
through the windowmanager classes (PhoneWM, WMPolicy, WMService) and
into RotationPolicy where it can be queried by the Settings app. This is
used by the Settings app to determine if it should show the "auto-rotate
screen" option or not.

Additionally, RotationPolicy.isRotationLockToggleSupported now returns
false if the device cannot detect orientation. This causes the Auto-rotate
option to not be shown in the notification bar on tablet devices.

Additionally, RotationPolicy.setRotationLock now takes a rotation
parameter to force the rotation to that value instead of the previous
USER_ROTATION value.

Change-Id: Ib7829d116f5aa08be4ba68438e155f47f9adb5b5
Signed-off-by: Matt Gumbel <matthew.k.gumbel@intel.com>