Get notification for laptop/tablet mode changes
For convertible tablet/laptops, add a notification so that
the framework knows when the device goes into or out of
tablet mode.

One possible usage is to disable certain input devices
when the device is in tablet mode.

Due to the current lack of interest in convertible devices
on AOSP, this patch is not recommended for upstreaming.

Category: Device Enablement
Domain: IO
Upstream-Candidate: no

Change-Id: I5e6c8de22bf099729494c8a94f9b528a93f9b39e
Signed-off-by: Daniel Leung <daniel.leung@intel.com>

Conflicts:

	services/java/com/android/server/wm/InputMonitor.java