Ability to disable input devices due to tablet mode changes
Adds the ability to disable certain input devices when going into
or out of tablet mode.

This fixes an issue where the touchpad is too sensitive and giving
phantom input events, due to the back of screen 'barely' touching it.

Due to the current lack of interest in convertible devices
on AOSP, this patch is not recommended for upstreaming.

Category: Device Enablement
Domain: IO
Upstream-Candidate: no

Change-Id: I497fea885d30d5f96eed1d2b63669991d38e37fa
Signed-off-by: Daniel Leung <daniel.leung@intel.com>

Conflicts:

	core/res/res/values/arrays.xml
	core/res/res/values/symbols.xml