Add power off button to quick settings dropdown box
Add an option to power off the device in the quick settings
drop-down dialog. It is necessary for devices which the power
button does not function as Android expected. For example,
the hardware only sends button down event but no button up
event, or there is not long press event emitted what-so-ever.
Clicking on this button asks user for confirmation, and fires
the shutdown intent.

Issue: AXIA-1271
Change-Id: I12c3af70d39d45a2974f8fca03eb332e68017e15
Original-Change-Id: Iba14b10d12e788f9df6038e20aa98384838531e0
Original-Change-Id: Ic973ebf43b79b436a9e872613b8572a7c77ce0c3
Signed-off-by: Daniel Leung <daniel.leung@intel.com>
Signed-off-by: Anton Cherkashyn <antonx.t.cherkashyn@intel.com>