Fix race condition with package manager and listeners
Observed in the android.content CTS test, but could happen anywhere.

The package is modified (which causes a broadcast intent to be sent
to all listeners), but the package is removed before the listeners
actually process the intent.  When they do process it, they don't
catch the exception that the package doesn't exist, causing them
to crash.

Issue: AXIA-2963
Change-Id: I3440d0613b5161b657fad4953297d264988f348c