BootReceiver: Add support to store kernel panic data in dropbox
Author:    Michael Bergeron <michael.bergeron@intel.com>
Author:    Robert Chiras <robert.chiras@intel.com>

This patch adds support for the aggregation of kernel panic files from
the pstore in to the Android drop box.
() Concatenates dmesg files for each kernel panic in to a single file
    and stores it in the Drop Box (/data/system/dropbox)
() Because they are binary and will need some extra parsing anyway,
    Android logs use the existing interfaces, and are stored as
    individual files.

The main purpose of this patch is to simplify internal debugging of
kernel panics, so upstreaming is not recommended at this time.

Category: AOSP Enhancement
Domain: Debugging
Upstream-Candidate: no

Issue: AXIA-3123
For: APDEV-736
Change-Id: Iaefdd135e8fd1bedd70568825538cf462e0d0ee0
Depends-Change-Id: I9abfc13d12ff16e10e4ab56119c8613c214da1b3
Signed-off-by: Michael Bergeron <michael.bergeron@intel.com>

Conflicts:

	services/java/com/android/server/BootReceiver.java