HW keyboard is not functional in Pin view
Pin view for screen lock can accept key events, the issue is focus is
not on the Pin view.
Modify pinentry to a EditText, so that it can always try to get focus.
And let security view request focus when resumed.

Due to the current lack of interest in laptop-type devices
on AOSP, this patch is not recommended for upstreaming.

Category: Device Enablement
Domain: IO
Upstream-Candidate: no

Issue: AXIA-2836
Change-Id: I94bc363542485bce47274f95442840e35404e20b
Signed-off-by: juntingwang <Junting.Wang@windriver.com>

Conflicts:

	core/res/res/layout/keyguard_pin_view.xml
	policy/src/com/android/internal/policy/impl/keyguard/KeyguardHostView.java