Assure StorageNotification passes correct volume for formatting
If a format request comes to the ExternalStorageFormatter, it will check the intent
for the extra 'volume' data. If not present, it will use the default external storage
path (/storage/emulated/legacy). This causes an IllegalArgumentException and a soft reboot.
The change assures that StorageNotification also sends in the intent the volume to be
formatted (eg. /storage/usbcard) for which the notification was received in the first place.

Issue: AXIA-3300
Category: device enablement
Domain: AOSP-Storage
Origin: internal
Upstream-Candidate: yes

Change-Id: I7c5c602f6473a227934faa619f5ab39c80454698
Signed-off-by: Adriana Reus <adriana.reus@intel.com>