Decoupled ADB debugging manager from UsbDeviceManager
Till now, the ADB debugging manager (also named UsbDebuggingManager) is
created from UsbDeviceManager since the debuging on most Android devices
is made through USB cable.
But, the adbd proess also listens on an IP port (which can be configured
or the default port 5555 can be used).
If we enable the secure feature in adbd, the back-end of this feature is
UsbDebuggingManager which handles the authorization of the connecting
host. But, UsbDebuggingManager only handles USB connections.
So, since adbd also listens on an IP port, we need to decouple
UsbDebuggingManger from USB and make it generic: DebuggingManager. Thus,
we will be able to make use of the secure feature in adbd also when
using it over IP, not just over USB.

This patch, does the followings:
- removes UsbDebuggingManager.java and creates a new service:
DebuggingService.java. This new service does exactly the same thing as
the old UsbDebuggingManager.java, but it will run as a system service,
having a permanent connection with adbd.
- The adb debugging controls are removed from UsbService and are put
into a new interface: IDebuggingManager, implemented by
DebuggingService.
- SystemServer is modified to add the new service: DebuggingService
- applications like SystemUI and Settings will be modify to use the new
service DebuggingService instead of UsbService when handling ADB debug
options (like: enable/disable and authorize). Settings will be updated
in a separate patch.

Change-Id: Ib37fbe07cb5f52a4138ca15040d415ecf839beb0
Depends-Change-Id: I40255a968ef52149bef93a0a38a68a6548f0ba21
Depends-Change-Id: I869643fe1200e6a6309ad1f9f7a130880eb7981e
Depends-Change-Id: I126b627d14807be52232b023d0230ef299cfd02e
Tracked-On: https://jira01.devtools.intel.com/browse/GMINL-4623
For: APDEV-896
Category: AOSP Improvement
Domain: AOSP-debugging
Origin: internal
Upstream-Candidate: Yes
Signed-off-by: Robert Chiras <robert.chiras@intel.com>
Signed-off-by: Petre Pircalabu <petre.ovidiu.pircalabu@intel.com>
Reviewed-on: https://android.intel.com/333881
Reviewed-by: Popa, Valentin <valentin.popa@intel.com>
Reviewed-by: Serban, Mihai <mihai.serban@intel.com>