Fix for battery status UNKNOWN sittuations
PowerUI:
When battery level is low, use the batteryPresent property instead of
batteryStatus to decide if the battery is draining in order to diplay
the low and critical level warnings.
BatteryService:
When battery level is 0, do not assume that it is charging if the
battery status is UNKNOWN. Checking if a charger is plugged is enough.

Change-Id: Ia4dc756d7c17f951383f904e403d35ed76eebdee
For: APDEV-943
Category: AOSP Improvement
Domain: AOSP-framework
Origin: internal
Upstream-Candidate: Yes
Signed-off-by: Robert Chiras <robert.chiras@intel.com>