Let isProviderEnabled default to the Settings.Scure value
isProviderEnabled must deafult to the values stored in Settings.Secure.
Those values are independent of the avaiable location providers.
The user can set a provider as enabled in Settings even if that provider
is not present. This creates an inconsistency in the APIs; Thus, I
have eliminated the extra check that LocationManagerService performs
and let it read directly from Settings.Secure

Issue: APDEV-884
Change-Id: I3bf72d52d9172e4baa0dece00d09088e6a7de1df
Category: AOSP Improvement
Domain: AOSP-framework
Origin: internal
Upstream-Candidate: Yes
Signed-off-by: Bogdan George Stefan <bogdan.g.stefan@intel.com>
Reviewed-on: https://android.intel.com/160138
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>
Reviewed-by: Breana, Tiberiu A <tiberiu.a.breana@intel.com>
Reviewed-by: Tomaszewski, Piotr <piotr.tomaszewski@intel.com>
Reviewed-by: Chiras, Robert <robert.chiras@intel.com>