Modify notification text for USB storage
When removing an external media check if the volume
description matches USB and sets the correct title
and message for the notification. If the resource
doesn't exist the default google texts will be used.

This patch needs to be re-worked with 7800bd8 in
packages/apps/Settings in order to upstream.

For: APDEV-1902
Category: aosp improvement
Domain: AOSP-Storage
Origin: internal
Upstream-Candidate: no

Change-Id: I583219f84f9c831884062074a5de7d695384271e
Signed-off-by: Markus Grape <markus.grape@intel.com>
Reviewed-on: https://android.intel.com/166957
Tested-by: sys_abtbuild <sys_abtbuild@intel.com>
Reviewed-by: Fagerstedt, Axel <axel.fagerstedt@intel.com>