Use TOUCHSCREEN_FINGER res configuration if mouse is detected
Platforms that don't have any touchscreen display will use
TOUCHSCREEN_NOTOUCH as res configuration by default.
But if no hardware MENU button is present on the device,
and if a specific app (ex. Gallery2) filters the action bar based
on 'notouch' res filter showing it only if the display is a
touchscreen one, then the user will not be able to access it.

This patch use TOUCHSCREEN_FINGER for res configuration if mouse is
detected, allowing user to interract with the soft MENU button using the
mouse.

Issue: APDEV-2828
Category: AOSP Improvement
Domain: AOSP-framework
Origin: internal
Upstream-Candidate: Yes

Change-Id: Ia66ff52cfdb89280ff949b6d9ef9d1e954706630
Signed-off-by: Rares Petrescu <rares.petrescu@intel.com>
Reviewed-on: https://android.intel.com/183689
Reviewed-by: Breana, Tiberiu A <tiberiu.a.breana@intel.com>
Reviewed-by: Musca, Constantin <constantin.musca@intel.com>
Reviewed-by: Chiras, Robert <robert.chiras@intel.com>